backend/requirements.txt
Flask==1.1.2
Flask_Cors==3.0.10
Flask_PyMongo==2.3.0
Flask_Session==0.4.0
keras==2.10.0
numpy==1.22.3
pandas==1.4.3
pymongo==3.12.3
python_bcrypt==0.3.2
scikit_learn==1.1.3
Werkzeug==1.0.1
yahoo_fin==0.8.9.1
